Add link to sponsor images fixes #374
Added the column imagelink to Sponsors, it's nullable since some
sponsors might not want to have a link.  If the string in the input
field is set to whitespace/nothing, then the value will be set to nil.
If the link is not a valid http(s) url, an error will be generated.